Move Expression code into public interface
